# Force use of LibYAML for yaml files parsing