--- a/dockerfiles/Dockerfile.test
+++ b/dockerfiles/Dockerfile.test
@@ -5,6 +5,7 @@ ENV LANG=en_US.utf8
 
 RUN yum install -y epel-release && \
     yum install -y python34 python36 && \
+    yum install -y python36-PyYAML && \
     yum clean all && \
     pip3 install --upgrade pip && \
     pip3 install tox