--- a/validator/bundler.py
+++ b/validator/bundler.py
@@ -8,7 +8,7 @@ import logging
 
 # Monkey-patch PyYAML forcing it to use LibYAML (i.e.: CLoader, CDumper)
 # Has to load before anymarkup
-import utils.pylibyaml 
+import utils.pylibyaml  # noqa: F401
 
 import anymarkup
 import click