--- a/validator/test/fixtures.py
+++ b/validator/test/fixtures.py
@@ -2,7 +2,7 @@ import os
 
 # Monkey-patch PyYAML forcing it to use LibYAML (i.e.: CLoader, CDumper)
 # Has to load before anymarkup
-import utils.pylibyaml 
+import utils.pylibyaml  # noqa: F401
 
 import anymarkup