--- a/validator/validator.py
+++ b/validator/validator.py
@@ -8,7 +8,7 @@ from functools import lru_cache
 
 # Monkey-patch PyYAML forcing it to use LibYAML (i.e.: CLoader, CDumper)
 # Has to load before anymarkup
-import utils.pylibyaml
+import utils.pylibyaml  # noqa: F401
 
 import anymarkup
 import click