# Context
For a long time I've been seeing qontract-validator take longer and longer to run. After digging a bit, this appears to be priparily due to the fact we're using a version of PyYAML that is not compiled with LibYAML support. Furthermore, it is not possible to tell anymarkdown to use the C-based LibYAML loaders and dumpers even if they are available.

# PyYAML
We list PyYAML as a dependency. This makes sense as it is required to run the app. However in order to ensure it is built with LibYAML support, the client also needs to have the libyaml and python development libraries in addition to having a compiler and related build tools.

In order to achieve this for our needs, I've updated the Dockerfiles to install PyYAML from the repos. Python finds the system installed package before the user-installed one (it searches /usr/lib64 before /usr/local/lib64)

# pylibyaml
I have debated between a few approaches as to how to change the code to use PyYAML with LibYAML support:
- Remove the anymarkup dependency and maintain our own json/yaml loaders (not exactly trivial)
- Use pylibyaml (https://github.com/philsphicas/pylibyaml) to monkey-patch / override PyYAML forcing it to use LibYAML
- Send a PR to anymarkup (planning to do but takes time & might not get accepted upstream)

I initially went with adding pylibyaml as a dependency but then decided to copy/vendor it into the codebase since it is small enough. I was uneasy with adding it as a dependency since that project is new, has a single author, no community, no stars, no forks and we do not know what the author will do with it in the future.

I realize this is a hack, but all things considerer this appears to be the less intrusive until (if) anymarkup supports LibYAML or we refactor the code to drop the anymarkup dependency

# Tests

The test suite is still reporting success after this change

Some rudimentary performance testing has shown the following results:
| Test   | Before   | After  | Improvement |
| -------|----------|--------|-------------|
| local  | 10s      | 1.5s    | 85%         |
| docker | 20s       | 5s      | 75%         |
| podman | 20s       | 4s      | 85%         |