@jfchevrette 2021-01-11
I spiked the idea of dropping anymarker completely, but I got a bit complicated when I realized anymarkup has some special configs for how yaml is parsed:
https://github.com/bkabrda/anymarkup-core/blob/08896a8215819edcc413e3f25885180469a4c2ed/anymarkup_core/__init__.py#L466-L495