@apahim 2021-01-11
Such a beautiful hack that utils/pylibyaml.py :heart:
Love the commits btw :slightly_smiling_face:
The places we use anymarkup seem a bit fine to be deterministic in terms of what format to expect there. The effort seems not that much and it pays off when we consider that this implementation is tightly coupled to the pyyaml source code.
So, my reccomendation is to try to use yaml with CLoader directly, if possible.
If that proves to be an issue, then you already have a working plan. Wdyt?