@rporres 2021-01-12
@apahim Do you mean to try drop anymarkup and use directly yaml with CLoader or json parsers instead where it is used as it is what we expect anyway, e.g. https://github.com/app-sre/qontract-validator/blob/master/validator/bundler.py#L36-L44?