@rporres 2021-01-12
I do agree with that, and I think @jfchevrette agrees too :)