@jfchevrette 2021-01-12
Less hacks version: https://github.com/app-sre/qontract-validator/pull/33